[#23] refactor: Chart를 컴포넌트로 분리
AsyncBoundary 적용으로 로딩과 에러 상태를 선언적으로 관리함
