+++ src/components/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { default as AsyncBoundary } from './AsyncBoundary';
+export { default as Chart } from './Chart';
